Remove usage of old grid inmplementation
